Refactor: restructure LLM client service with new  pattern and remove legacy code
