+++ backend/services/llm_client/__init__.py
@@ -1,0 +1,18 @@
+"""
+LLM Client Adapter Service
+
+Provides adapter interfaces and implementations for LLM providers.
+"""
+
+from backend.services.llm_client.base import LLMClient, LLMResponse
+from backend.services.llm_client.openai_client import OpenAILLMClient
+from backend.services.llm_client.development_client import DevelopmentLLMClient
+from backend.services.llm_client.factory import create_llm_client
+
+__all__ = [
+    "LLMClient",
+    "LLMResponse",
+    "OpenAILLMClient",
+    "DevelopmentLLMClient",
+    "create_llm_client",
+]
